Make LombScargle.power() respect units; add tests
